flask/buildimage/requirements.txt
flask
scikit-learn
plotly
pandas
kafka-python
Flask-SocketIO
eventlet
flask_cors
